Ensure Inner Simulation Reverts (#947)
This PR makes a small modification the `simulate` function such that it
ensures its inner `simulateAndRevert` reverts as expected. This is to
prevent calls to the `CompatibilityFallbackHandler` from making
unexpected state changes when not called by a Safe.

While using the `CompatibilityFallbackHandler` for something other than
a Safe fallback handler is explictely not supported, it makes sense to
ensure the inner call inverts and prevent unintended behaviours.