Ackee[M1]: Deprecate `createProxyWithCallback` Function (#955)
It is possible to front-run a `createProxyWithCallback` deployment using
`createProxyWithNonce` with a crafted `saltNonce`. This would make it so
the proxy would be deployed, but the callback not executed. Since this
method is not in use, we are deciding to deprecate it.

Alternatively, it would be possible to bring it back by using a
different `CREATE2` salt computation method in case this method is
needed in the future. Integrations that require a specific call at
deployment call can use the `setup` parameters with a `to` and `data`,
which would guarantee a callback at deployment time.