Ackee[W3]: Clear High `masterCopy` Bits (#962)
This PR changes the `SafeProxy` to write to the Solidity zero slot
instead of the scratch space. This ensures that the leading bits to the
encoded `masterCopy` value is guaranteed to be zero. While they
currently are 0 for all versions of Solidity that the code supports
(0.7.6 -> 0.8.30), this makes the code future-proof to potential
Compiler changes that may make the scratch space start dirty before
entering the `fallback` method.