Ackee[W6]: Prevent `TokenCallbackHandler` from Locking Tokens (#965)
This PR changes the implementation of `TokenCallbackHandler` and
`TokenCallbacks` to make a best-effort check on whether the tokens were
sent to a Safe that configures a token callback handler contract as a
fallback handler, or if it sent to the fallback handler contract itself
directly.

Thanks @remedcu and @rmeissner for the suggestions on how to implement
it! Basically, it works calling the `msg.sender` and expecting it to be
a Safe configured with the fallback handler. This would either:

- Fail if the token doesn't implement `StorageAccessible`
- Fail if the token doesn't have the specific `FALLBACK_HANDLER_SLOT`
set to the actual fallback handler

Note, however, that it is possible to trick the fallback handler into
receiving tokens. Because of this, we document that the checks are
best-effort only! We add a test to document how the fallback handler can
be tricked.