Prevent Self-Ownership for Non-EIP-7702 Safes (#1062)
Fixed #1055 

This PR changes the Safe smart account logic to disallow `owner == this`
in **unless** `this` is an EIP-7702 delegated account. This enables us
to use Safe as a delegation target, but without leaving in the footgun
where users can accidentally add a non-EIP-7702 Safe as its own owner
and effectively reducing the owner count (since a contract-based Safe
cannot sign for itself) and allowing situations where it is not possible
for generate a threshold of sigantures.

Additionally, I moved some Solidity files around to better reflect the
project layout:
- `Enum.sol` goes to the `interfaces/` directory, since it is arguably
an interface type
- `ErrorMessage.sol` goes into `common/` as it is not a library in the
Safe core contracts meaning of the word.

Added tests around the new logic, and with the refactoring I did in the
`OwnerManager.sol`, the code size actually decreased with this PR, even
with the added feature :)

As a side note, the EIP-7702 tests are unfortunately disabled for
coverage, as they fail with:

```
ProviderError: Cannot send both gasPrice and authorizationList
```

---------

Co-authored-by: Shebin John <admin@remedcu.com>